[chore] #57 Divide HomeScreen component in sub directory
